fix the function usage in the parser
